Changed method getAssigneeId to getUserIdentifier
getUserIdentifier is more generic and works for any user.
The change is used for the user assigned to the ticket, the user that requested the ticket, and any user in comments.
The initial priority comment is now assigned to the default user.
The UpdateAssignee method from TicketServiceImpl now also checks the stored value for changes.